Structure of samples changed to more modern setup.
