+++ Controls/LoadablePanel/src/DotVVM.Contrib.Samples/DotVVM.Contrib.Samples.csproj
@@ -14,15 +14,12 @@
 	<ItemGroup>
 		<None Remove="Controls\Pager\Pager.js" />
 		<None Remove="dotvvm_serialized_config.json.tmp" />
-		<None Remove="Views\PagingRepeater.dothtml" />
-		<None Remove="Views\Sample2.dothtml" />
-		<None Remove="Views\Sample3.dothtml" />
 	</ItemGroup>
 	<ItemGroup>
 	  <Content Include="Controls\Pager\Pager.js">
 	    <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
 	  </Content>
-	  <Content Include="Views\_master.dotmaster" />
+	  <Content Include="Pages\_master.dotmaster" />
 	</ItemGroup>
 	<ItemGroup>
 		<PackageReference Include="Microsoft.AspNetCore" Version="2.2.0" />
