+++ Controls/LoadablePanel/src/DotVVM.Contrib.Samples/Pages/Default/_default.dothtml
@@ -1,5 +1,5 @@
-﻿@viewModel DotVVM.Contrib.Samples.ViewModels.DefaultViewModel, DotVVM.Contrib.Samples
-@masterPage Views/_master.dotmaster
+﻿@viewModel DotVVM.Contrib.Samples.Pages.Default.DefaultViewModel, DotVVM.Contrib.Samples
+@masterPage Pages/_master.dotmaster
 
 <dot:Content ContentPlaceHolderID="MainContent">
