+++ Controls/LoadablePanel/src/DotVVM.Contrib.Samples/Pages/PagingRepeater/ItemModel.cs
@@ -1,0 +1,9 @@
+﻿namespace DotVVM.Contrib.Samples.Pages.PagingRepeater
+{
+    public class ItemModel
+    {
+        public string Id { get; set; }
+        public string Data { get; set; }
+    }
+}
+
