+++ Controls/LoadablePanel/src/DotVVM.Contrib.Samples/Pages/PagingRepeater/PagingRepeater.dothtml
@@ -1,17 +1,17 @@
-﻿@viewModel DotVVM.Contrib.Samples.ViewModels.PagingRepeaterViewModel, DotVVM.Contrib.Samples
-@masterPage Views/_master.dotmaster
-@service viewModel = DotVVM.Contrib.Samples.ViewModels.PagingRepeaterViewModel
+﻿@viewModel DotVVM.Contrib.Samples.Pages.PagingRepeater.PagingRepeaterViewModel, DotVVM.Contrib.Samples
+@masterPage Pages/_master.dotmaster
+@service service = DotVVM.Contrib.Samples.Services.GridService
 
 <dot:Content ContentPlaceHolderID="MainContent">
     <dot:RequiredResource Name="pager-js" />
-    <cc:Pager DataContext="{value: Grid}" Loader="{staticCommand: () => _root.Grid = viewModel.FillGrid(_root.Grid) }" />
+    <cc:Pager DataContext="{value: Grid}" Loader="{staticCommand: () => _root.Grid = service.FillGrid(_root.Grid)}" />
     <dot:Repeater DataSource="{value: _root.LoadingItems}">
          {{value: _this}}
     </dot:Repeater>
     <dot:Repeater class="repeater" DataSource="{value: Grid}">
         <dc:LoadablePanel ContentReloadBinding="{value: Id}" 
                           LoadingElementIds="{value: _root.LoadingItems}" 
-                          Load="{staticCommand: Data = viewModel.LoadItem(_this.Id).Result }" 
+                          Load="{staticCommand: Data = service.LoadItem(_this.Id).Result}" 
                           HideUntilLoaded="true" class="panel">
             <ProgressTemplate>
                 <div class="update-progress"><div class="background"></div><div class="spinner"><div class="rect1"></div><div class="rect2"></div><div class="rect3"></div><div class="rect4"></div><div class="rect5"></div></div></div>
@@ -19,7 +19,7 @@
             <ContentTemplate>
                 <div class="box">
                     <h1>Loaded {{value: Data}}</h1>
-                    <dot:Button Click="{command: _root.LoadItem(Id)}" Text="PostBack" />
+                    <dot:Button Click="{command: service.LoadItem(Id) }" Text="PostBack" />
                 </div>
             </ContentTemplate>
         </dc:LoadablePanel>
