+++ Controls/LoadablePanel/src/DotVVM.Contrib.Samples/Pages/PagingRepeater/PagingRepeaterViewModel.cs
@@ -1,0 +1,39 @@
+﻿using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+using DotVVM.Contrib.Samples.Services;
+using DotVVM.Framework.Controls;
+
+namespace DotVVM.Contrib.Samples.Pages.PagingRepeater
+{
+    public class PagingRepeaterViewModel : MasterViewModel
+    {
+        private readonly GridService itemLoadingService;
+
+        public GridViewDataSet<ItemModel> Grid { get; set; } = new GridViewDataSet<ItemModel>
+        {
+            PagingOptions = new PagingOptions
+            {
+                PageSize = 6,
+                TotalItemsCount = 32
+            }
+        };
+
+        public PagingRepeaterViewModel(GridService itemLoadingService)
+        {
+            this.itemLoadingService = itemLoadingService;
+        }
+
+        public List<string> LoadingItems { get; set; } = new List<string>();
+
+
+
+
+        public override Task PreRender()
+        {
+            itemLoadingService.FillGrid(Grid);
+            return base.PreRender();
+        }
+    }
+}
+
