+++ Controls/LoadablePanel/src/DotVVM.Contrib.Samples/Pages/Sample1/Sample1.dothtml
@@ -1,17 +1,17 @@
-﻿@viewModel DotVVM.Contrib.Samples.ViewModels.Sample1ViewModel, DotVVM.Contrib.Samples
-@masterPage Views/_master.dotmaster
-@service viewModel = DotVVM.Contrib.Samples.ViewModels.Sample1ViewModel
+﻿@viewModel DotVVM.Contrib.Samples.Pages.Sample1.Sample1ViewModel, DotVVM.Contrib.Samples
+@masterPage Pages/_master.dotmaster
+@service service = DotVVM.Contrib.Samples.Services.DataService
 
 <dot:Content ContentPlaceHolderID="MainContent">
 
-    <dc:LoadablePanel Load="{staticCommand: Data1 = viewModel.LoadData1().Result}" HideUntilLoaded="true" class="panel">
+    <dc:LoadablePanel Load="{staticCommand: Data1 = service.LoadData1().Result}" HideUntilLoaded="true" class="panel">
         <ContentTemplate>
             {{value: Data1}}<br />
             HideUntilLoaded="true"
         </ContentTemplate>
     </dc:LoadablePanel>
 
-    <dc:LoadablePanel Load="{staticCommand: Data2 = viewModel.LoadData2().Result}" HideUntilLoaded="false" class="panel">
+    <dc:LoadablePanel Load="{staticCommand: Data2 = service.LoadData2().Result}" HideUntilLoaded="false" class="panel">
         <ProgressTemplate>
             <div class="update-progress"><div class="background"></div><div class="spinner"><div class="rect1"></div><div class="rect2"></div><div class="rect3"></div><div class="rect4"></div><div class="rect5"></div></div></div>
         </ProgressTemplate>
