+++ Controls/LoadablePanel/src/DotVVM.Contrib.Samples/Pages/Sample1/Sample1ViewModel.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace DotVVM.Contrib.Samples.Pages.Sample1
+{
+    public class Sample1ViewModel : MasterViewModel
+    {
+        public string Data1 { get; set; }
+        public string Data2 { get; set; }
+    }
+}
+
