+++ Controls/LoadablePanel/src/DotVVM.Contrib.Samples/Pages/Sample2/Sample2.dothtml
@@ -1,10 +1,10 @@
-﻿@viewModel DotVVM.Contrib.Samples.ViewModels.Sample2ViewModel, DotVVM.Contrib.Samples
-@masterPage Views/_master.dotmaster
-@service viewModel = DotVVM.Contrib.Samples.ViewModels.Sample2ViewModel
+﻿@viewModel DotVVM.Contrib.Samples.Pages.Sample2.Sample2ViewModel, DotVVM.Contrib.Samples
+@masterPage Pages/_master.dotmaster
+@service service = DotVVM.Contrib.Samples.Services.ColorService
 
 <dot:Content ContentPlaceHolderID="MainContent">
     <div class="repeater">
-        <dc:LoadablePanel Load="{staticCommand: Data1 = viewModel.LoadData1().Result}" HideUntilLoaded="true" class="panel">
+        <dc:LoadablePanel Load="{staticCommand: Data1 = service.LoadData1().Result }" HideUntilLoaded="true" class="panel">
             <ProgressTemplate>
                 <div class="update-progress"><div class="background"></div><div class="spinner"><div class="rect1"></div><div class="rect2"></div><div class="rect3"></div><div class="rect4"></div><div class="rect5"></div></div></div>
             </ProgressTemplate>
@@ -13,7 +13,7 @@
             </ContentTemplate>
         </dc:LoadablePanel>
 
-        <dc:LoadablePanel Load="{staticCommand: Data2 = viewModel.LoadData2().Result}" HideUntilLoaded="true" class="panel">
+        <dc:LoadablePanel Load="{staticCommand: Data2 = service.LoadData2().Result}" HideUntilLoaded="true" class="panel">
             <ProgressTemplate>
                 <div class="update-progress"><div class="background"></div><div class="spinner"><div class="rect1"></div><div class="rect2"></div><div class="rect3"></div><div class="rect4"></div><div class="rect5"></div></div></div>
             </ProgressTemplate>
@@ -22,7 +22,7 @@
             </ContentTemplate>
         </dc:LoadablePanel>
 
-        <dc:LoadablePanel Load="{staticCommand: Data3 = viewModel.LoadData3().Result}" HideUntilLoaded="true" class="panel">
+        <dc:LoadablePanel Load="{staticCommand: Data3 = service.LoadData3().Result}" HideUntilLoaded="true" class="panel">
             <ProgressTemplate>
                 <div class="update-progress"><div class="background"></div><div class="spinner"><div class="rect1"></div><div class="rect2"></div><div class="rect3"></div><div class="rect4"></div><div class="rect5"></div></div></div>
             </ProgressTemplate>
@@ -31,7 +31,7 @@
             </ContentTemplate>
         </dc:LoadablePanel>
 
-        <dc:LoadablePanel Load="{staticCommand: Data4 = viewModel.LoadData4().Result}" HideUntilLoaded="true" class="panel">
+        <dc:LoadablePanel Load="{staticCommand: Data4 = service.LoadData4().Result}" HideUntilLoaded="true" class="panel">
             <ProgressTemplate>
                 <div class="update-progress"><div class="background"></div><div class="spinner"><div class="rect1"></div><div class="rect2"></div><div class="rect3"></div><div class="rect4"></div><div class="rect5"></div></div></div>
             </ProgressTemplate>
@@ -40,7 +40,7 @@
             </ContentTemplate>
         </dc:LoadablePanel>
 
-        <dc:LoadablePanel Load="{staticCommand: Data5 = viewModel.LoadData5().Result}" HideUntilLoaded="true" class="panel">
+        <dc:LoadablePanel Load="{staticCommand: Data5 = service.LoadData5().Result}" HideUntilLoaded="true" class="panel">
             <ProgressTemplate>
                 <div class="update-progress"><div class="background"></div><div class="spinner"><div class="rect1"></div><div class="rect2"></div><div class="rect3"></div><div class="rect4"></div><div class="rect5"></div></div></div>
             </ProgressTemplate>
@@ -49,7 +49,7 @@
             </ContentTemplate>
         </dc:LoadablePanel>
 
-        <dc:LoadablePanel Load="{staticCommand: Data6 = viewModel.LoadData6().Result}" HideUntilLoaded="true" class="panel">
+        <dc:LoadablePanel Load="{staticCommand: Data6 = service.LoadData6().Result}" HideUntilLoaded="true" class="panel">
             <ProgressTemplate>
                 <div class="update-progress"><div class="background"></div><div class="spinner"><div class="rect1"></div><div class="rect2"></div><div class="rect3"></div><div class="rect4"></div><div class="rect5"></div></div></div>
             </ProgressTemplate>
@@ -58,7 +58,7 @@
             </ContentTemplate>
         </dc:LoadablePanel>
 
-        <dc:LoadablePanel Load="{staticCommand: Data7 = viewModel.LoadData7().Result}" HideUntilLoaded="true" class="panel">
+        <dc:LoadablePanel Load="{staticCommand: Data7 = service.LoadData7().Result}" HideUntilLoaded="true" class="panel">
             <ProgressTemplate>
                 <div class="update-progress"><div class="background"></div><div class="spinner"><div class="rect1"></div><div class="rect2"></div><div class="rect3"></div><div class="rect4"></div><div class="rect5"></div></div></div>
             </ProgressTemplate>
@@ -67,7 +67,7 @@
             </ContentTemplate>
         </dc:LoadablePanel>
 
-        <dc:LoadablePanel Load="{staticCommand: Data8 = viewModel.LoadData8().Result}" HideUntilLoaded="true" class="panel">
+        <dc:LoadablePanel Load="{staticCommand: Data8 = service.LoadData8().Result}" HideUntilLoaded="true" class="panel">
             <ProgressTemplate>
                 <div class="update-progress"><div class="background"></div><div class="spinner"><div class="rect1"></div><div class="rect2"></div><div class="rect3"></div><div class="rect4"></div><div class="rect5"></div></div></div>
             </ProgressTemplate>
