+++ Controls/LoadablePanel/src/DotVVM.Contrib.Samples/Pages/Sample2/Sample2ViewModel.cs
@@ -1,0 +1,19 @@
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace DotVVM.Contrib.Samples.Pages.Sample2
+{
+    public class Sample2ViewModel : MasterViewModel
+    {
+        public string Data1 { get; set; }
+        public string Data2 { get; set; }
+        public string Data3 { get; set; }
+        public string Data4 { get; set; }
+        public string Data5 { get; set; }
+        public string Data6 { get; set; }
+        public string Data7 { get; set; }
+        public string Data8 { get; set; }
+    }
+}
+
