+++ Controls/LoadablePanel/src/DotVVM.Contrib.Samples/Pages/Sample3/Sample3.dothtml
@@ -1,6 +1,6 @@
-﻿@viewModel DotVVM.Contrib.Samples.ViewModels.Sample3ViewModel, DotVVM.Contrib.Samples
-@masterPage Views/_master.dotmaster
-@service viewModel = DotVVM.Contrib.Samples.ViewModels.Sample3ViewModel
+﻿@viewModel DotVVM.Contrib.Samples.Pages.Sample3.Sample3ViewModel, DotVVM.Contrib.Samples
+@masterPage Pages/_master.dotmaster
+@service service = DotVVM.Contrib.Samples.Services.ItemService
 
 <dot:Content ContentPlaceHolderID="MainContent">
     <dot:RequiredResource Name="globalize" />
@@ -11,14 +11,14 @@
     </div>
 
     <dot:Repeater class="repeater" DataSource="{value: Items}">
-        <dc:LoadablePanel Load="{staticCommand: _this.Data = viewModel.LoadItem(_this.Id).Result}" LoadingElementIds="{value: _root.LoadingItems}" HideUntilLoaded="true" class="panel">
+        <dc:LoadablePanel Load="{staticCommand: _this.Data = service.LoadItem(_this.Id).Result}" LoadingElementIds="{value: _root.LoadingItems}" HideUntilLoaded="true" class="panel">
             <ProgressTemplate>
                 <div class="update-progress"><div class="background"></div><div class="spinner"><div class="rect1"></div><div class="rect2"></div><div class="rect3"></div><div class="rect4"></div><div class="rect5"></div></div></div>
             </ProgressTemplate>
             <ContentTemplate>
                 <div class="box">
                     <h1>Loaded</h1>
-                    <dot:Button Click="{command: _root.LoadItem(Id)}" Text="PostBack" />
+                    <dot:Button Click="{command: service.LoadItem(Id)}" Text="PostBack" />
                 </div>
             </ContentTemplate>
         </dc:LoadablePanel>
