+++ Controls/LoadablePanel/src/DotVVM.Contrib.Samples/Services/ColorService.cs
@@ -1,23 +1,11 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System;
 using System.Threading.Tasks;
 using DotVVM.Framework.ViewModel;
 
-namespace DotVVM.Contrib.Samples.ViewModels
+namespace DotVVM.Contrib.Samples.Services
 {
-	public class Sample2ViewModel : MasterViewModel
-	{
-        public string Data1 { get; set; }
-        public string Data2 { get; set; }
-        public string Data3 { get; set; }
-        public string Data4 { get; set; }
-        public string Data5 { get; set; }
-        public string Data6 { get; set; }
-        public string Data7 { get; set; }
-        public string Data8 { get; set; }
-
+    public class ColorService
+    {
         [AllowStaticCommand]
         public async Task<string> LoadData1()
         {
