+++ Controls/LoadablePanel/src/DotVVM.Contrib.Samples/Services/DataService.cs
@@ -1,17 +1,10 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using DotVVM.Framework.ViewModel;
 
-namespace DotVVM.Contrib.Samples.ViewModels
+namespace DotVVM.Contrib.Samples.Services
 {
-	public class Sample1ViewModel : MasterViewModel
-	{
-        public string Data1 { get; set; }
-        public string Data2 { get; set; }
-
+    public class DataService
+    {
         [AllowStaticCommand]
         public async Task<string> LoadData1()
         {
