+++ Controls/LoadablePanel/src/DotVVM.Contrib.Samples/Services/ItemService.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Threading.Tasks;
+using DotVVM.Framework.ViewModel;
+
+namespace DotVVM.Contrib.Samples.Services
+{
+    public class ItemService
+    {
+        [AllowStaticCommand]
+        public async Task<string> LoadItem(string id)
+        {
+            await Task.Delay(GetRandomDelay());
+            return "black";
+        }
+
+        private int GetRandomDelay()
+        {
+            return new Random((int)DateTime.Now.Ticks).Next(200, 5000);
+        }
+    }
+}
+
